requirements.txt
imageio==2.37.2
numpy==2.0.2
opencv-python==4.13.0.92
pillow==11.3.0
tqdm==4.67.3
